extract ScanSourceOpDesc.scala for base
