+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/csv/CSVScanSourceOpDesc.scala
@@ -4,16 +4,14 @@ import com.fasterxml.jackson.annotation.{JsonIgnore, JsonProperty, JsonPropertyD
 import com.google.common.io.Files
 import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaTitle
 import edu.uci.ics.amber.engine.common.Constants
-import edu.uci.ics.texera.web.resource.dashboard.file.UserFileUtils
-import edu.uci.ics.texera.workflow.common.WorkflowContext
 import edu.uci.ics.texera.workflow.common.metadata.{
   OperatorGroupConstants,
   OperatorInfo,
   OutputPort
 }
-import edu.uci.ics.texera.workflow.common.operators.source.SourceOperatorDescriptor
 import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, Schema}
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.inferSchemaFromRows
+import edu.uci.ics.texera.workflow.operators.source.scan.ScanSourceOpDesc
 import org.codehaus.jackson.map.annotate.JsonDeserialize
 
 import java.io.{BufferedReader, File, FileReader, IOException}
@@ -22,15 +20,10 @@ import java.util.Collections.singletonList
 import scala.collection.JavaConverters._
 import scala.collection.immutable.List
 
-class CSVScanSourceOpDesc extends SourceOperatorDescriptor {
+class CSVScanSourceOpDesc extends ScanSourceOpDesc {
 
   @JsonIgnore
-  val INFER_READ_LIMIT: Int = 100
-
-  @JsonProperty(required = true)
-  @JsonSchemaTitle("File")
-  @JsonDeserialize(contentAs = classOf[java.lang.String])
-  var fileName: Option[String] = None
+  var headerLine: String = _
 
   @JsonProperty(defaultValue = ",")
   @JsonSchemaTitle("Delimiter")
@@ -43,9 +36,6 @@ class CSVScanSourceOpDesc extends SourceOperatorDescriptor {
   @JsonPropertyDescription("whether the CSV file contains a header line")
   var hasHeader: Boolean = true
 
-  @JsonIgnore
-  var filePath: Option[String] = None
-
   @throws[IOException]
   override def operatorExecutor: CSVScanSourceOpExecConfig = {
     // fill in default values
@@ -54,15 +44,14 @@ class CSVScanSourceOpDesc extends SourceOperatorDescriptor {
 
     filePath match {
       case Some(path) =>
-        val headerLine: String =
-          Files.asCharSource(new File(path), Charset.defaultCharset).readFirstLine
+        headerLine = Files.asCharSource(new File(path), Charset.defaultCharset).readFirstLine
 
         new CSVScanSourceOpExecConfig(
-          this.operatorIdentifier,
+          operatorIdentifier,
           Constants.defaultNumWorkers,
           path,
+          inferSchema(),
           delimiter.get.charAt(0),
-          this.inferSchema(headerLine),
           hasHeader
         )
       case None =>
@@ -81,37 +70,12 @@ class CSVScanSourceOpDesc extends SourceOperatorDescriptor {
     )
   }
 
-  @throws[IOException]
-  override def sourceSchema(): Schema = {
-    if (filePath.isEmpty) return null
-
-    val headerLine: String =
-      Files.asCharSource(new File(filePath.get), Charset.defaultCharset).readFirstLine
-    inferSchema(headerLine)
-
-  }
-
-  override def setContext(workflowContext: WorkflowContext): Unit = {
-    super.setContext(workflowContext)
-
-    if (context.userID.isDefined)
-      // if context has a valid user ID, the fileName will be a file name,
-      // resolve fileName to be the actual file path.
-      filePath = Option(
-        UserFileUtils.getFilePath(context.userID.get.toString, fileName.get).toString
-      )
-    else
-      // otherwise, the fileName will be inputted by user, which is the filePath.
-      filePath = fileName
-
-  }
-
   /**
     * Infer Texera.Schema based on the top few lines of data.
-    * @param headerLine usually the first line of the CSV file which contains table headers.
     * @return Texera.Schema build for this operator
     */
-  private def inferSchema(headerLine: String): Schema = {
+  @Override
+  def inferSchema(): Schema = {
     if (delimiter.isEmpty) return null
 
     val headers: Array[String] = headerLine.split(delimiter.get)
